+++ mobile/src/src/components/Login/index.jsx
@@ -1,0 +1,3 @@
+import LoginPage from "./LoginPage";
+
+export default LoginPage;
