+++ mobile/src/src/components/Register/index.jsx
@@ -1,0 +1,3 @@
+import RegistrationPage from "./RegistrationPage";
+
+export default RegistrationPage;
